Move getWorkdirForRepoUrl to utils
